Add wasm32-unknown-emscripten support
Convert overflow-prone constants (FALLBACK_MEMORY_BYTES, MAX_INVERTED_MINIMIZERS,
MAX_INVERTED_BUCKET_IDS, MAX_CHUNK_BYTES) from usize to u64 so they compile on
wasm32 where usize is 32-bit. Add .min(usize::MAX as u64) as usize capping at
all runtime boundaries. Change format_bytes() to accept u64, remove duplicate in
merge.rs. Fix parse_byte_suffix f64 precision bug (> to >=). Guard
advise_prefetch with cfg(not(target_arch = "wasm32")) since emscripten sets
cfg(unix). Fix parse_slurm_memory and ONE_TB constants to use u64. Gate 64-GiB
budget tests with cfg(target_pointer_width = "64"). Add wasm32-check CI job and
docs/WASM.md.

+++ src/commands/classify.rs
@@ -114,8 +114,8 @@ pub fn run_classify(args: ClassifyRunArgs) -> Result<()> {
             "Adaptive batch sizing: batch_size={}, threads={}, estimated peak memory={}, shard_reservation={}, format={:?}",
             batch_result.batch_size,
             rayon::current_num_threads(),
-            format_bytes(batch_result.peak_memory),
-            format_bytes(batch_result.shard_reservation),
+            format_bytes(batch_result.peak_memory as u64),
+            format_bytes(batch_result.shard_reservation as u64),
             batch_result.input_format
         );
     }
@@ -596,8 +596,8 @@ pub fn run_log_ratio(args: ClassifyLogRatioArgs) -> Result<()> {
             "Adaptive batch sizing: batch_size={}, threads={}, estimated peak memory={}, shard_reservation={}, format={:?}",
             batch_result.batch_size,
             rayon::current_num_threads(),
-            format_bytes(batch_result.peak_memory),
-            format_bytes(batch_result.shard_reservation),
+            format_bytes(batch_result.peak_memory as u64),
+            format_bytes(batch_result.shard_reservation as u64),
             batch_result.input_format
         );
     }
@@ -834,7 +834,7 @@ pub fn run_log_ratio(args: ClassifyLogRatioArgs) -> Result<()> {
         log::info!(
             "Flushing {} remaining deferred reads (~{}) after final batch (batch {})",
             deferred_buffer.len(),
-            format_bytes(deferred_buffer.approx_bytes()),
+            format_bytes(deferred_buffer.approx_bytes() as u64),
             batch_num
         );
         let t_drain = std::time::Instant::now();
@@ -1178,7 +1178,7 @@ fn process_log_ratio_batch<S: AsRef<str>>(
         log::info!(
             "Deferred buffer reached threshold ({} reads, ~{}), flushing (triggered by batch {})",
             deferred_buffer.len(),
-            format_bytes(deferred_buffer.approx_bytes()),
+            format_bytes(deferred_buffer.approx_bytes() as u64),
             batch_num
         );
         let t_drain = std::time::Instant::now();
